Allow "export html foo.zip" to provide name non-interactively.
I have it set right now to not overwrite any existing file with that name.

Currently it saves into the TIC data directory rather than the current
directory; if you pass it an absolute path it fails; I believe the
path should be interpreted as relative to the current directory, but
I'm not sure how to accomplish this.