Update to Fennel version 0.6.0.
Rename modified compiler file to loadfennel to make it clearer that
it's not the exact same thing as the upstream module. (It has a shim
in it to populate package.loaded because it will not be loaded
normally thru require.)

Remove mention of fennelfriend module; it is now integrated with the
main compiler.

Add some keywords to syntax highlighting.