Copy off src field of CommandDesc before tokenizing command string.
The CommandDesc struct has two relevant fields here: src and
command. The src field is supposed to contain the source of the entire
input while the command field is supposed to contain just the name of
the command being invoked; aka everything before the first space.

We would start with input like this:

    eval trace("hello" .. "world")

What we want is:

   src: eval trace("hello" .. "world")
   command: eval

The src field was set to a copy of the input, and then the command
field was set to running strtok on src to find the first space. But
even though strtok returns what looks like a string, it's actually
just pointing to the same copy of src which now happens to have been
terminated right where the first space used to be.

So what we got was just:

   src: eval
   command: eval

The fix is to copy the input into src and command fields separately
and only call strtok on command, allowing src to be left alone.

Finally since src and command are now separate strings we must free
them separately rather than only freeing src.