Fix the check for strict mode in Fennel headers.
This makes it so that unknown globals need to use the _G table; this
feature was originally added in a5c443b.

It was broken in 61ffe39 when a spurious space was added to the
headers. Technically this is a breaking change, but hopefully it won't
affect too many carts.